+++ doma-core/src/main/java/org/seasar/doma/jdbc/query/AutoBatchUpsertQuery.java
@@ -35,7 +35,7 @@ public class AutoBatchUpsertQuery<ENTITY> extends AutoBatchInsertQuery<ENTITY>
     Dialect dialect = config.getDialect();
     PreparedSqlBuilder builder = new PreparedSqlBuilder(config, SqlKind.BATCH_UPSERT, sqlLogType);
     UpsertContext context =
-        UpsertContext.fromEntity(
+        UpsertContextBuilder.fromEntity(
             builder,
             entityType,
             duplicateKeyType,
